made order view

+++ database/migrations/2019_04_09_071842_create_product_category_table.php
@@ -14,9 +14,9 @@ class CreateProductCategoryTable extends Migration
     public function up()
     {
         Schema::create('category_product', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('product_id');
-            $table->unsignedBigInteger('category_id');
+            $table->increments('id');
+            $table->integer('product_id');
+            $table->integer('category_id');
             $table->timestamps();            
         });
